Commit all work done

+++ ulyp-ui/src/main/kotlin/com/ulyp/ui/RecordingTreeNodeContent.kt
@@ -5,7 +5,7 @@ import javafx.geometry.Pos
 import javafx.scene.layout.StackPane
 
 // TODO should be a better name
-class RecordingTreeNodeContent(node: CallRecord, renderSettings: RenderSettings?, totalNodeCountInTree: Long) : StackPane() {
+class RecordingTreeNodeContent(node: CallRecord, renderSettings: RenderSettings?, totalNodeCountInTree: Int) : StackPane() {
 
     init {
         alignment = Pos.CENTER_LEFT
